pyproject.toml
[project]
name = "raillock"
version = "0.1.0"
description = "A Python library and CLI for MCP tool validation and review."
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "mcp==1.5.0",
    "requests",
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/yourusername/raillock"

[tool.setuptools]
packages = ["raillock"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "pexpect",
    "mcp==1.5.0",
    "trio",
    "vulture",
    "tabulate"
]

[project.scripts]
raillock = "raillock.cli.__main__:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta" 